General update and add p10k and gitconfig.
New MacBook Pro refresh of these config files. No longer using 'Fig' so
I've removed the references to it.

I deleted the bashrc file as I don't use that on any system. The zshrc
file is fully based on oh-my-zsh *and* Powerlevel10k. I haven't tested
the auto-install of p10k yet so that may fail on actual use. (That 'git
clone' command can be removed if needed.)

Tmux config is not complete but is in working state and I like where
it's at.